+++ src/tables/mod.rs
@@ -33,6 +33,7 @@ pub mod os2;
 pub mod post;
 pub mod sbix;
 pub mod stat;
+pub mod svg;
 pub mod vhea;
 pub mod vmtx;
 pub mod vorg;